// src/app/models/user.ts
export class User{ 
    // lezem export bech tetchef 
    idCustomer: number; 
     //derniere version angular te5alik lezem initialisi value donc temchi tebadel regle fil ts.config fil strict = false
    firstName: string;
    lastName: string;
    birthDate: string;
    email: string;
    password: string;
    profession: string;
    accountCategory: string;
    picture: string;
    }
    